+++ source/write_output/Write_Stoichiometry_Matrix_For_Opt.cpp
@@ -10,7 +10,6 @@
 void Write_Stoichiometric_Matrix_For_Opt
 (
 		string filename ,
-		//const vector< string >& Species,
 		const vector< SingleReactionData >& Reactions
 )
 {
@@ -22,9 +21,9 @@ void Write_Stoichiometric_Matrix_For_Opt
 	{
 
 		int Number_Reactions = 0;
-		Number_Reactions = Reactions.size();
+		Number_Reactions = (int)Reactions.size();
 		int Number_Species = 0;
-		Number_Species = Reactions[0].Reactants.size();
+		Number_Species = (int)Reactions[0].Reactants.size();
 
 		Output << "   ";
